+++ NEWS.md
@@ -6,7 +6,9 @@ Version 1.1.1 - beta
 --------------------
 - Release notes: https://ezdxf.mozman.at/release-v1-1.html
 - NEW: Python 3.12 binary wheel deployment on PyPI
-- NEW: page alignment support in the `drawing` add-on for these backends: 
+- NEW: page alignment support for the `drawing` add-on for these backends: 
+  `SVGBackend`, `PyMuPdfBackend` and the `PlotterBackend`
+- NEW: cropping content at page margins for the `drawing` add-on for these backends:
   `SVGBackend`, `PyMuPdfBackend` and the `PlotterBackend`
 - NEW: support for decoding of MIF encoded text `\M+cxxxx` by the `recover` module
 - INFO: `numpy` v1.25 has stopped providing Python 3.8 binary wheels on PyPI